MyApps-13: Move web implementation, api and dependency injection to submodules.
