+++ app/src/main/cpp/CMakeLists.txt
@@ -7,5 +7,4 @@ cmake_minimum_required(VERSION 3.22.1)
 
 add_subdirectory(basic)
 add_subdirectory(myengine)
-add_subdirectory(web)
 add_subdirectory(calculator)
