--- app/src/main/cpp/web/.gitignore
@@ -1,1 +1,0 @@
-*/api_key.h
