+++ app/src/main/java/xyz/torquato/myapps/di/data/DataModule.kt
@@ -7,12 +7,10 @@ import dagger.hilt.components.SingletonComponent
 import xyz.torquato.myapps.api.math.IMathRepository
 import xyz.torquato.myapps.api.web.ILocalWebRepository
 import xyz.torquato.myapps.api.web.IQueryRepository
-import xyz.torquato.myapps.api.web.IWebRepository
 import xyz.torquato.myapps.api.web.cache.IWebCacheRepository
 import xyz.torquato.myapps.data.math.MathRepository
 import xyz.torquato.myapps.data.web.LocalWebRepository
 import xyz.torquato.myapps.data.web.QueryRepository
-import xyz.torquato.myapps.data.web.WebRepository
 import xyz.torquato.myapps.data.web.cache.WebCacheRepository
 import javax.inject.Singleton
 
@@ -24,9 +22,7 @@ abstract class RepositoryModule {
     @Singleton
     abstract fun bindMathImpl(impl: MathRepository): IMathRepository
 
-    @Binds
-    @Singleton
-    abstract fun bindWebImpl(impl: WebRepository): IWebRepository
+
 
     @Binds
     @Singleton
