+++ app/src/main/java/xyz/torquato/myapps/domain/web/SetSelectedBookUseCase.kt
@@ -1,8 +1,6 @@
 package xyz.torquato.myapps.domain.web
 
 import xyz.torquato.myapps.api.web.ILocalWebRepository
-import xyz.torquato.myapps.api.web.IQueryRepository
-import xyz.torquato.myapps.api.web.model.QueryRequest
 
 class SetSelectedBookUseCase(
     private val repository: ILocalWebRepository
