+++ data/di/src/main/java/xyz/torquato/myapps/data/di/RepositoryImpl.kt
@@ -7,14 +7,20 @@ import dagger.hilt.components.SingletonComponent
 import xyz.torquato.myapps.data.impl.cypher.CypherRepository
 import xyz.torquato.myapps.data.impl.paging.PagingRepository
 import xyz.torquato.myapps.data.impl.waves.SoundRepository
+import xyz.torquato.myapps.data.impl.web.WebRepository
 import xyz.torquato.myaps.domain.api.encrypt.ICypherRepository
 import xyz.torquato.myaps.domain.api.pagging.IPagingRepository
 import xyz.torquato.myaps.domain.api.sound.ISoundRepository
+import xyz.torquato.myaps.domain.api.web.IWebRepository
 import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class RepositoryImpl {
+
+    @Binds
+    @Singleton
+    abstract fun bindWebImpl(impl: WebRepository): IWebRepository
 
     @Binds
     abstract fun bindImpl(impl: SoundRepository): ISoundRepository
