+++ data/impl/src/main/cpp/CMakeLists.txt
@@ -7,3 +7,4 @@ cmake_minimum_required(VERSION 3.22.1)
 
 add_subdirectory(cypher)
 add_subdirectory(waves)
+add_subdirectory(web)
