+++ data/impl/src/main/cpp/web/.gitignore
@@ -1,0 +1,1 @@
+secrets/api_key.h
