+++ data/impl/src/main/java/xyz/torquato/myapps/data/impl/web/WebDataSource.kt
@@ -1,6 +1,5 @@
-package xyz.torquato.myapps.data.web
+package xyz.torquato.myapps.data.impl.web
 
-import org.json.JSONObject
 import javax.inject.Inject
 import javax.inject.Singleton
