+++ domain/api/src/main/java/xyz/torquato/myaps/domain/api/web/model/BookError.kt
@@ -1,4 +1,4 @@
-package xyz.torquato.myapps.api.web.model
+package xyz.torquato.myaps.domain.api.web.model
 
 data class BookError(
     val message: String
