+++ domain/api/src/main/java/xyz/torquato/myaps/domain/api/web/model/QueryRequest.kt
@@ -1,4 +1,4 @@
-package xyz.torquato.myapps.api.web.model
+package xyz.torquato.myaps.domain.api.web.model
 
 data class QueryRequest(
     val query: String,
