+++ domain/api/src/main/java/xyz/torquato/myaps/domain/api/web/model/QueryResult.kt
@@ -1,4 +1,4 @@
-package xyz.torquato.myapps.api.web.model
+package xyz.torquato.myaps.domain.api.web.model
 
 sealed interface QueryResult {
